+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix: intercept any 401 error in a response
 - Upgrade frontend to Angular 16
 - Remove feature: Public IP
+- Make Admin menu in frontend a dropdown
 
 ## 2.23.6 (8 May 2023)
 - Fix regression: dropdown to change language